Clean up code related to io pipe Add env varibale to control cursor batch size
